Changed one more time to Integer

+++ quadratic/forms.py
@@ -1,9 +1,9 @@
 from django import forms
 
 class QuadraticForm(forms.Form):        
-    a = forms.FloatField(label='коэффициент a')
-    b = forms.FloatField(label='коэффициент b')
-    c = forms.FloatField(label='коэффициент c')
+    a = forms.IntegerField(label='коэффициент a')
+    b = forms.IntegerField(label='коэффициент b')
+    c = forms.IntegerField(label='коэффициент c')
     
     def clean_a(self):
         data = self.cleaned_data['a']
